resolve login and signuo

+++ src/database/query/post/deletePostQuery.js
@@ -1,5 +1,5 @@
 const { getQuery } = require('../getQuery');
 
-const deletePostQuery = (id) => getQuery(`DELETE FROM post WHERE id = "${id}" AND seen = 0 `, [id]);
+const deletePostQuery = (id, user_id) => getQuery(`DELETE FROM post WHERE id = "${id}" AND seen = 0 AND user_id = "${user_id}"  `, [id]);
 
 module.exports = deletePostQuery;
